scripts/requirements.txt
pandas
paho-mqtt
openpyxl